Change some SensorStateClass from MEASUREMENT to TOTAL_INCREASING
With this change the sensors

* Heutiger Strom-Export
* Heutiger Strom-Bezug
* Battereieentladung heute
* Batterieladung Heute

will change theier SensorStateClass to `TOTAL_INCREASING`
With this this sensors will be able to be put into the energy dashboard,
otherwise HA will complain about that these sensor with `measurement` does not have an `last_reset`.
This behaviour was changed a while back ago.